+++ src/components/Headline/Headline.mdx
@@ -1,5 +1,5 @@
 import { Playground, Props } from 'docz'
-import Headline from '@components/Headline'
+import { Headline } from '@components/Headline'
 
 # Headline
