+++ src/components/index.ts
@@ -35,7 +35,7 @@ export { default as FormikToggle } from './Form/Toggle/FormikToggle'
 export { default as Toggle } from './Form/Toggle/Toggle'
 export { default as GalleryImage } from './GalleryImage'
 export * from './Heading'
-export { default as Headline } from './Headline'
+export * from './Headline'
 export { default as Html } from './Html'
 export { default as Icon } from './Icon'
 export { default as Image } from './Image'
